+++ CHANGELOG.md
@@ -4,4 +4,5 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Unreleased
 
+* Add DateParam::parse for parsing relative date strings from features
 * Project commenced